Move Ram files into ram subdirectories
