+++ src/AstTranslator.cpp
@@ -20,15 +20,6 @@
 #include "FunctorOps.h"
 #include "Global.h"
 #include "LogStatement.h"
-#include "RamCondition.h"
-#include "RamExpression.h"
-#include "RamNode.h"
-#include "RamOperation.h"
-#include "RamProgram.h"
-#include "RamRelation.h"
-#include "RamStatement.h"
-#include "RamTranslationUnit.h"
-#include "RamUtils.h"
 #include "RelationTag.h"
 #include "SrcLocation.h"
 #include "ast/AstAbstract.h"
@@ -49,6 +40,15 @@
 #include "ast/analysis/AuxArityAnalysis.h"
 #include "ast/analysis/PrecedenceGraph.h"
 #include "json11.h"
+#include "ram/RamCondition.h"
+#include "ram/RamExpression.h"
+#include "ram/RamNode.h"
+#include "ram/RamOperation.h"
+#include "ram/RamProgram.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamTranslationUnit.h"
+#include "ram/RamUtils.h"
 #include "utility/ContainerUtil.h"
 #include "utility/FunctionalUtil.h"
 #include "utility/StreamUtil.h"
