+++ src/InterpreterEngine.cpp
@@ -26,19 +26,19 @@
 #include "InterpreterRelation.h"
 #include "Logger.h"
 #include "ProfileEvent.h"
-#include "RamCondition.h"
-#include "RamExpression.h"
-#include "RamOperation.h"
-#include "RamProgram.h"
-#include "RamRelation.h"
-#include "RamStatement.h"
 #include "RamTypes.h"
-#include "RamVisitor.h"
 #include "ReadStream.h"
 #include "RecordTable.h"
 #include "SignalHandler.h"
 #include "SymbolTable.h"
 #include "WriteStream.h"
+#include "ram/RamCondition.h"
+#include "ram/RamExpression.h"
+#include "ram/RamOperation.h"
+#include "ram/RamProgram.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamVisitor.h"
 #include "utility/EvaluatorUtil.h"
 #include "utility/MiscUtil.h"
 #include "utility/ParallelUtil.h"
