+++ src/InterpreterGenerator.h
@@ -22,17 +22,17 @@
 #include "InterpreterNode.h"
 #include "InterpreterPreamble.h"
 #include "InterpreterRelation.h"
-#include "RamCondition.h"
-#include "RamExpression.h"
-#include "RamIndexAnalysis.h"
-#include "RamNode.h"
-#include "RamOperation.h"
-#include "RamRelation.h"
-#include "RamStatement.h"
 #include "RamTypes.h"
-#include "RamUtils.h"
-#include "RamVisitor.h"
 #include "RelationTag.h"
+#include "ram/RamCondition.h"
+#include "ram/RamExpression.h"
+#include "ram/RamNode.h"
+#include "ram/RamOperation.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamUtils.h"
+#include "ram/RamVisitor.h"
+#include "ram/analysis/RamIndexAnalysis.h"
 #include "utility/MiscUtil.h"
 #include <algorithm>
 #include <cstddef>
