+++ src/InterpreterProgInterface.h
@@ -19,15 +19,15 @@
 #include "InterpreterEngine.h"
 #include "InterpreterIndex.h"
 #include "InterpreterRelation.h"
-#include "RamNode.h"
-#include "RamProgram.h"
-#include "RamRelation.h"
-#include "RamStatement.h"
-#include "RamTranslationUnit.h"
 #include "RamTypes.h"
-#include "RamVisitor.h"
 #include "SouffleInterface.h"
 #include "SymbolTable.h"
+#include "ram/RamNode.h"
+#include "ram/RamProgram.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamTranslationUnit.h"
+#include "ram/RamVisitor.h"
 #include <cassert>
 #include <cstddef>
 #include <cstdint>
