+++ src/Synthesiser.cpp
@@ -19,22 +19,22 @@
 #include "BinaryConstraintOps.h"
 #include "FunctorOps.h"
 #include "Global.h"
-#include "RamCondition.h"
-#include "RamExpression.h"
-#include "RamIndexAnalysis.h"
-#include "RamNode.h"
-#include "RamOperation.h"
-#include "RamProgram.h"
-#include "RamRelation.h"
-#include "RamStatement.h"
-#include "RamTranslationUnit.h"
 #include "RamTypes.h"
-#include "RamUtils.h"
-#include "RamVisitor.h"
 #include "RelationTag.h"
 #include "SymbolTable.h"
 #include "SynthesiserRelation.h"
 #include "json11.h"
+#include "ram/RamCondition.h"
+#include "ram/RamExpression.h"
+#include "ram/RamNode.h"
+#include "ram/RamOperation.h"
+#include "ram/RamProgram.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamTranslationUnit.h"
+#include "ram/RamUtils.h"
+#include "ram/RamVisitor.h"
+#include "ram/analysis/RamIndexAnalysis.h"
 #include "utility/FileUtil.h"
 #include "utility/MiscUtil.h"
 #include "utility/StreamUtil.h"
