+++ src/SynthesiserRelation.cpp
@@ -8,8 +8,8 @@
 
 #include "SynthesiserRelation.h"
 #include "Global.h"
-#include "RamIndexAnalysis.h"
 #include "RelationTag.h"
+#include "ram/analysis/RamIndexAnalysis.h"
 #include "utility/StreamUtil.h"
 #include <algorithm>
 #include <cassert>
