+++ src/main.cpp
@@ -22,11 +22,6 @@
 #include "InterpreterEngine.h"
 #include "InterpreterProgInterface.h"
 #include "ParserDriver.h"
-#include "RamNode.h"
-#include "RamProgram.h"
-#include "RamTransformer.h"
-#include "RamTransforms.h"
-#include "RamTranslationUnit.h"
 #include "RamTypes.h"
 #include "Synthesiser.h"
 #include "ast/AstNode.h"
@@ -41,6 +36,11 @@
 #include "ast/transform/ComponentInstantiationTransformer.h"
 #include "config.h"
 #include "profile/Tui.h"
+#include "ram/RamNode.h"
+#include "ram/RamProgram.h"
+#include "ram/RamTranslationUnit.h"
+#include "ram/transform/RamTransformer.h"
+#include "ram/transform/RamTransforms.h"
 #include "utility/FileUtil.h"
 #include "utility/StreamUtil.h"
 #include "utility/StringUtil.h"
