+++ src/ram/analysis/RamAnalysis.h
