+++ src/ram/analysis/RamComplexityAnalysis.cpp
@@ -14,12 +14,12 @@
  *
  ***********************************************************************/
 
-#include "RamComplexityAnalysis.h"
-#include "RamCondition.h"
-#include "RamExpression.h"
-#include "RamNode.h"
-#include "RamRelation.h"
-#include "RamVisitor.h"
+#include "ram/analysis/RamComplexityAnalysis.h"
+#include "ram/RamCondition.h"
+#include "ram/RamExpression.h"
+#include "ram/RamNode.h"
+#include "ram/RamRelation.h"
+#include "ram/RamVisitor.h"
 #include <cassert>
 
 namespace souffle {
