+++ src/ram/analysis/RamComplexityAnalysis.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "RamAnalysis.h"
+#include "ram/analysis/RamAnalysis.h"
 
 namespace souffle {
