+++ src/ram/transform/RamTransformer.cpp
@@ -14,14 +14,13 @@
  *
  ***********************************************************************/
 
-#include "RamTransformer.h"
+#include "ram/transform/RamTransformer.h"
 #include "DebugReport.h"
 #include "ErrorReport.h"
 #include "Global.h"
-#include "RamNode.h"
-#include "RamProgram.h"
-#include "RamTranslationUnit.h"
-#include "ast/transforms/DebugReporter.h"
+#include "ram/RamNode.h"
+#include "ram/RamProgram.h"
+#include "ram/RamTranslationUnit.h"
 #include <chrono>
 #include <cstdlib>
 #include <iostream>
