+++ src/ram/transform/RamTransforms.cpp
@@ -14,21 +14,21 @@
  *
  ***********************************************************************/
 
-#include "RamTransforms.h"
+#include "ram/transform/RamTransforms.h"
 #include "BinaryConstraintOps.h"
 #include "FunctorOps.h"
-#include "RamComplexityAnalysis.h"
-#include "RamCondition.h"
-#include "RamExpression.h"
-#include "RamNode.h"
-#include "RamOperation.h"
-#include "RamProgram.h"
-#include "RamRelation.h"
-#include "RamStatement.h"
 #include "RamTypes.h"
-#include "RamUtils.h"
-#include "RamVisitor.h"
 #include "RelationTag.h"
+#include "ram/RamCondition.h"
+#include "ram/RamExpression.h"
+#include "ram/RamNode.h"
+#include "ram/RamOperation.h"
+#include "ram/RamProgram.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamUtils.h"
+#include "ram/RamVisitor.h"
+#include "ram/analysis/RamComplexityAnalysis.h"
 #include "utility/ContainerUtil.h"
 #include "utility/MiscUtil.h"
 #include <algorithm>
