+++ src/tests/ram_arithmetic_test.cpp
@@ -21,14 +21,14 @@
 #include "FunctorOps.h"
 #include "Global.h"
 #include "InterpreterEngine.h"
-#include "RamExpression.h"
-#include "RamOperation.h"
-#include "RamProgram.h"
-#include "RamRelation.h"
-#include "RamStatement.h"
-#include "RamTranslationUnit.h"
 #include "RamTypes.h"
 #include "SymbolTable.h"
+#include "ram/RamExpression.h"
+#include "ram/RamOperation.h"
+#include "ram/RamProgram.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamTranslationUnit.h"
 #include <algorithm>
 #include <cmath>
 #include <cstddef>
