+++ src/tests/ram_expression_equal_clone_test.cpp
@@ -17,8 +17,8 @@
 #include "tests/test.h"
 
 #include "FunctorOps.h"
-#include "RamExpression.h"
 #include "RamTypes.h"
+#include "ram/RamExpression.h"
 #include <memory>
 #include <string>
 #include <utility>
