+++ src/tests/ram_operation_equal_clone_test.cpp
@@ -18,11 +18,11 @@
 
 #include "AggregateOp.h"
 #include "BinaryConstraintOps.h"
-#include "RamCondition.h"
-#include "RamExpression.h"
-#include "RamOperation.h"
-#include "RamRelation.h"
 #include "RelationTag.h"
+#include "ram/RamCondition.h"
+#include "ram/RamExpression.h"
+#include "ram/RamOperation.h"
+#include "ram/RamRelation.h"
 #include <memory>
 #include <string>
 #include <utility>
