+++ src/tests/ram_relation_equal_clone_test.cpp
@@ -16,8 +16,8 @@
 
 #include "tests/test.h"
 
-#include "RamRelation.h"
 #include "RelationTag.h"
+#include "ram/RamRelation.h"
 #include <string>
 
 namespace souffle {
