+++ src/tests/ram_relation_test.cpp
@@ -20,16 +20,16 @@
 #include "ErrorReport.h"
 #include "Global.h"
 #include "InterpreterEngine.h"
-#include "RamExpression.h"
-#include "RamOperation.h"
-#include "RamProgram.h"
-#include "RamRelation.h"
-#include "RamStatement.h"
-#include "RamTranslationUnit.h"
 #include "RamTypes.h"
 #include "RelationTag.h"
 #include "SymbolTable.h"
 #include "json11.h"
+#include "ram/RamExpression.h"
+#include "ram/RamOperation.h"
+#include "ram/RamProgram.h"
+#include "ram/RamRelation.h"
+#include "ram/RamStatement.h"
+#include "ram/RamTranslationUnit.h"
 #include <algorithm>
 #include <cstddef>
 #include <iostream>
